+++ mygpiod/lib/events.h
@@ -1,6 +1,6 @@
 /*
  SPDX-License-Identifier: GPL-3.0-or-later
- myGPIOd (c) 2020-2023 Juergen Mang <mail@jcgames.de>
+ myGPIOd (c) 2020-2024 Juergen Mang <mail@jcgames.de>
  https://github.com/jcorporation/myGPIOd
 */
 
@@ -16,7 +16,8 @@
 enum mygpiod_event_types {
     MYGPIOD_EVENT_FALLING,
     MYGPIOD_EVENT_RISING,
-    MYGPIOD_EVENT_LONG_PRESS
+    MYGPIOD_EVENT_LONG_PRESS,
+    MYGPIOD_EVENT_LONG_PRESS_RELEASE
 };
 
 /**